centralize shared function in utils.py and remove duplicate code
